refactor(router): extract routes/governance.rs (q1 wave 3/5)
Move AGT governance handlers + sensitive_agt_routes() builder into
routes/governance.rs:

  - agt_evaluate, agt_trust_{list,get,delete,update}
  - agt_audit, agt_audit_verify
  - agt_mesh_inbox (pub(super) — consumed by routes::mesh)
  - agt_status (uses mesh::lookup_parent_amid)
  - agt_signing_counter, agt_rate_limit_{get,update}
  - agt_reputation

Cross-module wiring:
  - mesh.rs now imports agt_mesh_inbox via super::governance
  - governance.rs imports lookup_parent_amid via super::mesh

Public API unchanged: routes::sensitive_agt_routes() still exported.

Proof: ZERO drift (85/85 items), 178/178 tests, clippy clean.
